redmine_interface.py: add support for additional trackers
This commit add support for multiple trackers
(ie ticket categories) through the use of the redmine section
in the bot's configuration file.
It works by reading the 'trackers' section of the config file
and dynamically mapping the configured trackers to those detected
by querying the redmine server (that is also configured in the bot's
config file).

The sample config file has been updated with an example
of how to configure the use of trackers.

redmine_interface_test.py updated with test code
for the support for additional trackers.

[TechnuxBot #12]

Change-Id: I7f9827365376a6a4b31d1f0e055267e4cdc22f25
Signed-off-by: Magnus Karlsson <magnus@technux.se>