ircbot.py: handle messages in threads and change config handling
Currently all received messages are handled in serial fashion,
ie. one message at a time. In order to avoid taking long time for users
to get their answers from the bot while it is handling for example
a redmine command message that may result in relatively lengthy communication
with a redmine server, threading is used to handle messages in the background
while still being able to handle new incoming messages.

To make the code more readable, the config handling is changed a bit.
All config values that are used outside the main function are to be
stored in a dict. The dict is supplied along with any other arguments
to the functions that need to access the config values.

While at it, make a minor clarification in the included config file.

Change-Id: I5e8328dead2bd0c277a9ff60c9ed6fd8945bd75a
Signed-off-by: Magnus Karlsson <magnus@technux.se>