Add sysvinit start script
Make it possible to use a sysvinit start script in order to automate the
startup of TechnuxBot on systems running sysvinit based startup system.

The default settings require that the code is located in
/opt/technuxbot/ and the configuration in /etc/technuxbot.conf

Special thanks to Magnus Karlsson <magnus@technux.se> for me being able to base
this script upon the work he did in
https://github.com/Technux/meta-technux.git - '9c991a Add recipe
for technuxbot - initial version'.

Signed-off-by: Petter Mabäcker <petter@technux.se>